abc/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
pydantic>=2.0.0
pytz>=2023.3
